requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
pymysql
SQLAlchemy
email_validator
psycopg2-binary
python-multipart